# internal/server/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Tracepad API",
    "summary": "Lightweight, self-hosted, OTLP-native store for LLM and agent traces.",
    "description": "The read API is the product's contract: the UI, the CLI and the MCP server are thin clients over exactly these endpoints, and a feature that is not here does not exist. Responses are flat snake_case JSON with RFC 3339 timestamps and opaque cursors. Every response spends a byte budget on payloads (default 51200, `?budget=` to override): a payload that does not fit is cut at a UTF-8 boundary and replaced by a truncation marker naming where the whole of it lives.",
    "license": {"name": "Apache-2.0", "identifier": "Apache-2.0"},
    "version": "0.1.0"
  },
  "servers": [{"url": "http://localhost:4318", "description": "A locally running tracepad"}],
  "security": [{"bearerKey": []}, {"basicKeys": []}],
...
  "paths": {
...
    "/api/v1/stats": {
      "get": {
        "tags": ["stats"],
        "summary": "Counts, errors, cost and latency percentiles per bucket",
        "description": "`unit` says what a bucket counts: grouping by hour, day, environment, release or total counts traces, and grouping by model counts observations, because a trace has no model. Grouped by release, a trace that named none falls in the bucket with the empty key. Grouped by `total`, the whole window is one bucket under the empty key — the same shape as a day bucket, with percentiles merged over every hour in the window — and it takes every filter the other groupings take. An empty range answers with no buckets rather than with fabricated zero rows. Latency percentiles are histogram-based: accurate to a few percent, and stable across the expiry of the raw rows.",
        "operationId": "stats",
        "parameters": [
          {"$ref": "#/components/parameters/From"},
          {"$ref": "#/components/parameters/To"},
          {"$ref": "#/components/parameters/Environment"},
          {"name": "user_id", "in": "query", "schema": {"type": "string"},
            "description": "Restricts every bucket to one end user. The shape, the groupings and `unit` are unchanged; a `hour` or `day` timeline, and the `total` bucket, additionally carry `sessions` per bucket. Statistics for one user trail the raw data by the same rollup lag as the rest"},
          {"name": "group_by", "in": "query",
            "schema": {"type": "string", "enum": ["hour", "day", "model", "environment", "release", "total"], "default": "day"}}
        ],
        "responses": {
          "200": {"description": "The buckets", "content": {"application/json": {"schema": {
            "type": "object",
            "properties": {
              "group_by": {"type": "string", "enum": ["hour", "day", "model", "environment", "release", "total"]},
              "unit": {"type": "string", "enum": ["trace", "observation"]},
              "buckets": {"type": "array", "items": {
                "type": "object",
                "properties": {
                  "key": {"type": "string"},
                  "count": {"type": "integer"},
                  "error_count": {"type": "integer"},
                  "total_cost": {"type": "number", "description": "Absent when nothing in the bucket carried a cost"},
                  "tokens": {"type": "object",
                    "description": "Token sums over the observations in the bucket, on every grouping and both units. Each key is present only when something in the bucket carried that count, and the object is absent when none of the three is — absent, never zero, like `total_cost` — and always absent with `user_id`, whose rollup holds no token sums. Cache-read tokens are the input tokens a provider reported as served from its cache; reasoning and cache-creation counts are not summed",
                    "properties": {
                      "input": {"type": "integer"},
                      "output": {"type": "integer"},
                      "cache_read": {"type": "integer"}
                    }},
                  "sessions": {"type": "integer",
                    "description": "Present only with `user_id` and an `hour`, `day` or `total` grouping: how many of that user's sessions began in this bucket. A session is counted where it starts, so a sum over any range is exact"},
                  "latency_ms": {"type": "object", "properties": {
                    "p50": {"type": ["integer", "null"]},
                    "p95": {"type": ["integer", "null"]}
                  }}
                },
                "required": ["key", "count", "error_count", "latency_ms"]
              }}
            },
            "required": ["group_by", "unit", "buckets"]
          }}}},
          "400": {"$ref": "#/components/responses/BadRequest"},
          "401": {"$ref": "#/components/responses/Unauthorized"},
          "503": {"$ref": "#/components/responses/ServiceUnavailable"}
        }
      }
    },
...
    "/api/v1/facets": {
      "get": {
        "tags": ["stats"],
        "summary": "The environments, releases and trace names in a range, each with its trace count",
        "description": "What the three many-valued filters can be set to, for the range in view: the distinct values of `environment`, `release` and `name` among the traces between `from` and `to`, each with the number of traces carrying it, sorted by count descending and then by value. It answers the range and nothing else — the counts do not respect the other filters, so the list does not change as boxes are ticked. At most 100 values per column; `omitted` says how many were left out, rarest first. A trace with no release is not a release, and a trace with no name is not a name. Answered from the hourly roll-up behind the watermark and from the raw rows for the tail, so a value first seen a minute ago is already here.",
        "operationId": "facets",
        "parameters": [
          {"name": "from", "in": "query", "schema": {"type": "string", "format": "date-time"},
            "description": "RFC 3339, inclusive. Defaults to the oldest hour the rollup holds, which is also as far back as this endpoint looks: it does not scan the raw traces for hours the rollup no longer covers. A window whose `from` is not before its `to` is a 400"},
          {"name": "to", "in": "query", "schema": {"type": "string", "format": "date-time"},
            "description": "RFC 3339, exclusive. Defaults to now"}
        ],
        "responses": {
          "200": {"description": "The values", "content": {"application/json": {"schema": {
            "type": "object",
            "properties": {
              "from": {"type": "string", "format": "date-time", "description": "The start of the range these counts are for"},
              "to": {"type": "string", "format": "date-time", "description": "Its exclusive end"},
              "environment": {"type": "array", "items": {"$ref": "#/components/schemas/FacetValue"}},
              "release": {"type": "array", "items": {"$ref": "#/components/schemas/FacetValue"}},
              "name": {"type": "array", "items": {"$ref": "#/components/schemas/FacetValue"}},
              "omitted": {"type": "object",
                "description": "How many values the cap left out of each column; 0 when the list is the whole truth",
                "properties": {
                  "environment": {"type": "integer"},
                  "release": {"type": "integer"},
                  "name": {"type": "integer"}
                },
                "required": ["environment", "release", "name"]}
            },
            "required": ["from", "to", "environment", "release", "name", "omitted"]
          }}}},
          "400": {"$ref": "#/components/responses/BadRequest"},
          "401": {"$ref": "#/components/responses/Unauthorized"},
          "503": {"$ref": "#/components/responses/ServiceUnavailable"}
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "bearerKey": {"type": "http", "scheme": "bearer",
        "description": "A project's secret key, `tp-sk-…`"},
      "basicKeys": {"type": "http", "scheme": "basic",
        "description": "`base64(public key:secret key)`, the shape a Langfuse SDK sends"},
      "adminToken": {"type": "http", "scheme": "bearer",
        "description": "`TRACEPAD_ADMIN_TOKEN`, sent in the same header as a project key. Required for everything cross-project, for creating and deleting a project, and for the account routes; beside an owner's or editor's session, it is what lists, mints and revokes a project's keys, which no project key does — which is what makes it the recovery when every owner's password is lost. A server with none configured answers 403 and says so."},
      "sessionCookie": {"type": "apiKey", "in": "cookie", "name": "tracepad_session",
        "description": "A browser session: `HttpOnly`, `SameSite=Lax`, `Secure` over TLS, set by the three ways in and lasting `TRACEPAD_SESSION_DAYS` (default 30), sliding. On a route whose policy is `member` or `editor` a session must also name the project it is asking about with `X-Tracepad-Project`, and an unsafe method must carry an `Origin` (or `Referer`) whose host is the request's `Host`, the first value of `X-Forwarded-Host`, or the host of `TRACEPAD_URL`."}
    },
    "parameters": {
      "From": {"name": "from", "in": "query", "schema": {"type": "string", "format": "date-time"},
        "description": "RFC 3339, inclusive. The range is half-open, so walking a timeline never reports a row twice."},
      "To": {"name": "to", "in": "query", "schema": {"type": "string", "format": "date-time"},
        "description": "RFC 3339, exclusive"},
      "Environment": {"name": "environment", "in": "query", "schema": {"type": "string"},
        "description": "The environment a trace ran in. A comma-separated list matches **any** of them: `?environment=production,staging`. Items are trimmed and duplicates collapse; an empty item (`a,,b`, `a,`) is a 400, as an empty value is, and so is repeating the parameter — one list, one parameter. At most 100 items; beyond that, a 400. An environment whose name contains a comma is not expressible here. `GET /api/v1/facets` lists the values in a range with their counts"},
...
    },
    "responses": {
      "BadRequest": {"description": "The request cannot mean what it says: an unknown query parameter, a malformed value, a limit out of range",
        "content": {"application/json": {"schema": {"$ref": "#/components/schemas/Error"}}}},
      "Unauthorized": {"description": "The credentials do not resolve to a project",
        "content": {"application/json": {"schema": {"$ref": "#/components/schemas/Error"}}}},
      "ServiceUnavailable": {"description": "The credentials could not be checked right now, more passwords were being checked at once than the server allows, or a write met a database condition that passes (a full disk, a lock that did not clear), or a read found no free read slot before its deadline (`the server is busy; retry shortly`): retry after `Retry-After`. A read the deadline stopped while it ran (`the read took longer than 20s and was stopped; narrow the time range or the filters`, the number being `TRACEPAD_READ_TIMEOUT`) carries no `Retry-After`: the same request would be stopped again, so narrow it",
        "headers": {"Retry-After": {"schema": {"type": "string"}, "description": "Seconds to wait"}},
        "content": {"application/json": {"schema": {"$ref": "#/components/schemas/Error"}}}},
...
    },
    "schemas": {
      "Error": {
        "type": "object",
        "description": "The one error shape the whole surface uses",
        "properties": {"error": {"type": "string"}},
        "required": ["error"]
      },
...
      "FacetValue": {
        "type": "object",
        "description": "One value a filter can be set to, and how many traces of the range carry it",
        "properties": {
          "value": {"type": "string", "description": "The value, exactly as `?environment=`, `?release=` or `?name=` takes it"},
          "count": {"type": "integer", "description": "How many traces of the range carry it"}
        },
        "required": ["value", "count"]
      },
...
    }
  }
}